// src/app/models/updateorder.model.ts
export interface EditOrder {
    OrderId: string;
    OrderDate: string;
    CreatedBy: string;
    CreatedOn: string;
    ApprovedRejectedBy: string;
    ApprovedRejectedOn: string;
    Quantity: string;
    AllowedOrder: string;
    Status: string;
    HospitalId: string;
    DepartmentId: string;
    ProductId: string;
    UserId: string
}